+++ src/plugin/Routes.php
@@ -78,7 +78,6 @@ trait Routes
             $event->rules['commerce/settings/lineitemstatuses/<id:\d+>'] = 'commerce/line-item-statuses/edit';
 
             // Store Settings
-
             $event->rules['commerce/store-settings/location'] = 'commerce/store-location/edit-location';
 
             $event->rules['commerce/store-settings/paymentcurrencies'] = 'commerce/payment-currencies/index';
@@ -88,7 +87,6 @@ trait Routes
             $event->rules['commerce/store-settings/donation'] = 'commerce/donations/edit';
 
             // Store Settings - Regions
-
             $event->rules['commerce/store-settings/countries'] = 'commerce/countries/index';
             $event->rules['commerce/store-settings/countries/new'] = 'commerce/countries/edit';
             $event->rules['commerce/store-settings/countries/<id:\d+>'] = 'commerce/countries/edit';
@@ -98,12 +96,10 @@ trait Routes
             $event->rules['commerce/store-settings/states/<id:\d+>'] = 'commerce/states/edit';
 
             // Lite shipping and tax
-
             $event->rules['commerce/store-settings/shipping'] = 'commerce/lite-shipping/edit';
             $event->rules['commerce/store-settings/tax'] = 'commerce/lite-tax/edit';
 
             // Shipping
-
             $event->rules['commerce/shipping/shippingzones'] = 'commerce/shipping-zones/index';
             $event->rules['commerce/shipping/shippingzones/new'] = 'commerce/shipping-zones/edit';
             $event->rules['commerce/shipping/shippingzones/<id:\d+>'] = 'commerce/shipping-zones/edit';