+++ src/services/Customers.php
@@ -712,9 +712,9 @@ class Customers extends Component
         }
 
         Craft::$app->getView()->registerAssetBundle(CommerceCpAsset::class);
-        return Craft::$app->getView()->renderTemplate('commerce/customers/_editUserTab', [
+        return Craft::$app->getView()->renderTemplate('commerce/customers/_includes/_editUserTab', [
             'customer' => $customer,
-            'userId' => $context['user']->id,
+            'addressRedirect' => $context['user']->getCpEditUrl(),
         ]);
     }
 }