+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Added the ability to enable/disabled countries. ([#213](https://github.com/craftcms/commerce/issues/213))
 - Added the ability to enable/disabled states. ([#213](https://github.com/craftcms/commerce/issues/213))
 - Added consolidation of guest orders after an order is completed. ([#1062](https://github.com/craftcms/commerce/issues/1062))
+- Added the ability to manager Customers and Customer Addresses. ([#1043](https://github.com/craftcms/commerce/issues/1043))
+- Added the ability to show the customer on the Order index table.
 - Added `craft\commerce\controllers\CountriesController::actionUpdateStatus()`
 - Added `craft\commerce\controllers\DiscountsController::actionClearDiscountUses()`
 - Added `craft\commerce\controllers\DiscountsController::actionUpdateStatus()`
@@ -25,7 +27,9 @@
 - Added `craft\commerce\controllers\SalesController::actionUpdateStatus()`
 - Added `craft\commerce\controllers\StatesController::actionUpdateStatus()`
 - Added `craft\commerce\elements\Order::getAdjustmentsByType()`.
+- Added `craft\commerce\elements\Order::getCustomerLinkHtml()`.
 - Added `craft\commerce\models\Country::$enabled`.
+- Added `craft\commerce\models\Customer::getCpEditUrl()`.
 - Added `craft\commerce\models\Discount::$totalDiscountUseLimit`.
 - Added `craft\commerce\models\Discount::$totalDiscountUses`.
 - Added `craft\commerce\models\ProductType::$titleLabel`.